# docs/swagger.yaml
basePath: /
definitions:
...
  db.PropertyMap:
    additionalProperties: true
    type: object
...
  db.TextSnippet:
    properties:
      date_created:
        type: string
      id:
        type: integer
      last_edited:
        type: string
      snippet:
        type: string
      title:
        type: string
      workspace_uuid:
        type: string
    type: object
...
  db.WfRequest:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: integer
      project_id:
        type: string
      request_data:
        $ref: '#/definitions/db.PropertyMap'
      request_id:
        type: string
      response_data:
        $ref: '#/definitions/db.PropertyMap'
      source:
        type: string
      status:
        $ref: '#/definitions/db.WfRequestStatus'
      updated_at:
        type: string
      workflow_id:
        type: string
    type: object
  db.WfRequestStatus:
    enum:
    - NEW
    - PENDING
    - COMPLETED
    - FAILED
    type: string
    x-enum-varnames:
    - StatusNew
    - StatusPending
    - StatusCompleted
    - StatusFailed
...
  handlers.CreateWorkflowRequestRequest:
    properties:
      request_id:
        type: string
      response_data:
        $ref: '#/definitions/db.PropertyMap'
    type: object
...
  handlers.SnippetRequest:
    properties:
      snippet:
        type: string
      title:
        type: string
    type: object
...
host: localhost:5002
info:
  contact: {}
  description: This is the API documentation for Sphinx Tribes.
  title: Sphinx Tribes API
  version: "1.0"
paths:
...
  /snippet/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a snippet by its ID
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Snippet deleted successfully
          schema:
            type: string
      summary: Delete Snippet
      tags:
      - Snippets
    get:
      consumes:
      - application/json
      description: Get a snippet by its ID
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TextSnippet'
      summary: Get Snippet by ID
      tags:
      - Snippets
    put:
      consumes:
      - application/json
      description: Update an existing snippet
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Snippet Request
        in: body
        name: snippet
        required: true
        schema:
          $ref: '#/definitions/handlers.SnippetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TextSnippet'
      summary: Update Snippet
      tags:
      - Snippets
  /snippet/create:
    post:
      consumes:
      - application/json
      description: Create a new snippet
      parameters:
      - description: Workspace UUID
        in: query
        name: workspace_uuid
        required: true
        type: string
      - description: Snippet Request
        in: body
        name: snippet
        required: true
        schema:
          $ref: '#/definitions/handlers.SnippetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.TextSnippet'
      summary: Create Snippet
      tags:
      - Snippets
  /snippet/workspace/{workspace_uuid}:
    get:
      consumes:
      - application/json
      description: Get snippets by workspace UUID
      parameters:
      - description: Workspace UUID
        in: path
        name: workspace_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.TextSnippet'
            type: array
      summary: Get Snippets by Workspace
      tags:
      - Snippets
...
  /workflows/request:
    post:
      consumes:
      - application/json
      description: Handle a workflow request
      parameters:
      - description: Workflow Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/db.WfRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle Workflow Request
      tags:
      - Workflows
  /workflows/response:
    post:
      consumes:
      - application/json
      description: Handle a workflow response
      parameters:
      - description: Workflow Response
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkflowRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle Workflow Response
      tags:
      - Workflows
...
swagger: "2.0"